imprówment

+++ src/main/java/com/project/cwmsgradle/controlls/ClientAddController.java
@@ -41,7 +41,7 @@ public class ClientAddController {
         // Używamy konstruktora z 4 argumentami
         Client newClient = new Client(name, surname, phone, email);
 
-        newClient.setClientId((long) clientId); // Ustawiamy clientId za pomocą settera
+        newClient.setClientId(clientId); // Ustawiamy clientId za pomocą settera
 
         clientMenageController.addClientToList(newClient);
         navigateToClientMenage(event);
