+++ src/main/java/com/project/cwmsgradle/utils/AuthenticatedUser.java
@@ -33,7 +33,7 @@ public class AuthenticatedUser {
         return client;
     }
 
-    public Long getClientId() {
+    public Integer getClientId() {
         return client != null ? client.getClientId() : null;
     }
 }
